Fixed benchmark_registry problem
The benchmark registry is now correct and each individual benchmark functions properly as does the multi-suite-capable benchmark.
